Introduce placeholders with conversions
Extend the regular placeholder syntax "${name}" by an optional
conversion part, like this: "${name|`conversion`}". The conversion is
evaluated as perl code, but somewhat safely wrapped using perl’s Safe
Compartment evaluation. See perldoc Safe for details.

If the pipe and the conversion are not specified, placeholders behave as
before. If a value is not defined, it is replaced by the empty string,
independent of whether there is a conversion or not.