Switch to MCE::Hobo for more control
This patch dumps our dependency to MCE::Flow and replaces the flow
mechanics with MCE::Hobo. MCE::Hobo lets us fork and join manually and
thus allows for more control over the processes.

Due to the fact that this is a pure refactoring, there shouldn't be any
effective changes due to this patch. It is supposed to be a preparatory
change for features like watchdogs, restarting of failed sources or
dynamically launched processes.