Add primitive watchdog
This patch adds an endless loop in the main process (that is used to
interpret the "config" and launch up all the needed child processes).
Inside the main loop, we implement some primitive watchdog. This first
implementation assumes, that no sub process is supposed to end. It
watches and regularly checks the processes and restarts died ones.
"Oneshot" or temporary processes are not supported yet. As of now, all
our built-in sources comply with that assumption.