hlwm plugin respects reload and quit_panel hooks
When hlwm emits the reload or quit_panel hooks, panels ought to exit,
because they will be restarted afterwards by hlwm, or hlwm terminates
and the panel is not needed anymore.

Since ulcer with the hlwm plugin is likely the driver for the panel, the
plugin now respects those hooks and signals ulcer’s main process to
shutdown.