Add support for "sections" in template string
This patch introduces "sections" to dissect a template string into
understandable parts. The user defines sections in the config file and
uses them in the template string or in other sections.

We ignore the danger of endless recursive patterns, because our users
are not total idiots (haha, only I am using this shit anyway, that’s the
joke). The behaviour is described in the feature section though.

The expansion of section references happens only once. There’s no need
to repeat it, because it’s all static.